Move Resource-specific functionality to own module
There was a bunch of functionality that was not actually specific
to LitDatasets, but also applied to non-RDF Resources (e.g.
related to fetching its ACL). These are now moved into a separate
`resource` module.
