+++ src/litDataset.ts
@@ -20,12 +20,9 @@
  */
 
 import { Quad, NamedNode } from "rdf-js";
-import LinkHeader from "http-link-header";
 import { dataset, DataFactory } from "./rdfjs";
-import { fetch } from "./fetcher";
 import { turtleToTriples, triplesToTurtle } from "./formats/turtle";
 import { isLocalNode, resolveIriForLocalNodes } from "./datatypes";
-import { internal_fetchResourceAcl, internal_fetchFallbackAcl } from "./acl";
 import {
   UrlString,
   LitDataset,
@@ -35,11 +32,13 @@ import {
   hasResourceInfo,
   LocalNode,
   unstable_WithAcl,
-  unstable_hasAccessibleAcl,
-  unstable_Access,
-  unstable_AclDataset,
-  unstable_WithAccessibleAcl,
 } from "./interfaces";
+import {
+  internal_parseResourceInfo,
+  internal_defaultFetchOptions,
+  internal_fetchAcl,
+  getFetchedFrom,
+} from "./resource";
 
 /**
  * Initialise a new [[LitDataset]] in memory.
@@ -51,12 +50,6 @@ export function createLitDataset(): LitDataset {
 }
 
 /**
- * @internal
- */
-export const defaultFetchOptions = {
-  fetch: fetch,
-};
-/**
  * Fetch a LitDataset from the given URL. Currently requires the LitDataset to be available as [Turtle](https://www.w3.org/TR/turtle/).
  *
  * @param url URL to fetch a [[LitDataset]] from.
@@ -65,10 +58,12 @@ export const defaultFetchOptions = {
  */
 export async function fetchLitDataset(
   url: UrlString,
-  options: Partial<typeof defaultFetchOptions> = defaultFetchOptions
+  options: Partial<
+    typeof internal_defaultFetchOptions
+  > = internal_defaultFetchOptions
 ): Promise<LitDataset & WithResourceInfo> {
   const config = {
-    ...defaultFetchOptions,
+    ...internal_defaultFetchOptions,
     ...options,
   };
 
@@ -83,167 +78,12 @@ export async function fetchLitDataset(
   const resource = dataset();
   triples.forEach((triple) => resource.add(triple));
 
-  const resourceInfo = parseResourceInfo(response);
+  const resourceInfo = internal_parseResourceInfo(response);
 
   const resourceWithResourceInfo: LitDataset &
     WithResourceInfo = Object.assign(resource, { resourceInfo: resourceInfo });
 
   return resourceWithResourceInfo;
-}
-
-/**
- * Retrieve the information about a resource (e.g. access permissions) without
- * fetching the resource itself.
- *
- * @param url URL to fetch Resource metadata from.
- * @param options Optional parameter `options.fetch`: An alternative `fetch` function to make the HTTP request, compatible with the browser-native [fetch API](https://developer.mozilla.org/en-US/docs/Web/API/WindowOrWorkerGlobalScope/fetch#Parameters).
- * @returns Promise resolving to the metadata describing the given Resource, or rejecting if fetching it failed.
- */
-export async function internal_fetchResourceInfo(
-  url: UrlString,
-  options: Partial<typeof defaultFetchOptions> = defaultFetchOptions
-): Promise<WithResourceInfo["resourceInfo"]> {
-  const config = {
-    ...defaultFetchOptions,
-    ...options,
-  };
-
-  const response = await config.fetch(url, { method: "HEAD" });
-  if (!response.ok) {
-    throw new Error(
-      `Fetching the Resource metadata failed: ${response.status} ${response.statusText}.`
-    );
-  }
-
-  const resourceInfo = parseResourceInfo(response);
-
-  return resourceInfo;
-}
-
-/**
- * This (currently internal) function fetches the ACL indicated in the [[WithResourceInfo]]
- * attached to a resource.
- *
- * @internal
- * @param resourceInfo The Resource info with the ACL URL
- * @param options Optional parameter `options.fetch`: An alternative `fetch` function to make the HTTP request, compatible with the browser-native [fetch API](https://developer.mozilla.org/en-US/docs/Web/API/WindowOrWorkerGlobalScope/fetch#Parameters).
- */
-export async function internal_fetchAcl(
-  resourceInfo: WithResourceInfo,
-  options: Partial<typeof defaultFetchOptions> = defaultFetchOptions
-): Promise<unstable_WithAcl["acl"]> {
-  if (!unstable_hasAccessibleAcl(resourceInfo)) {
-    return {
-      resourceAcl: null,
-      fallbackAcl: null,
-    };
-  }
-  const [resourceAcl, fallbackAcl] = await Promise.all([
-    internal_fetchResourceAcl(resourceInfo, options),
-    internal_fetchFallbackAcl(resourceInfo, options),
-  ]);
-
-  return {
-    fallbackAcl: fallbackAcl,
-    resourceAcl: resourceAcl,
-  };
-}
-
-/**
- * Experimental: fetch a LitDataset and its associated Access Control List.
- *
- * This is an experimental function that fetches both a Resource's metadata, the linked ACL Resource (if
- * available), and the ACL that applies to it if the linked ACL Resource is not available (if accessible). This can
- * result in many HTTP requests being executed, in lieu of the Solid spec mandating servers to
- * provide this info in a single request. Therefore, and because this function is still
- * experimental, prefer [[fetchLitDataset]] instead.
- *
- * If the Resource's linked ACL Resource could not be fetched (because it does not exist, or because
- * the authenticated user does not have access to it), `acl.resourceAcl` will be `null`. If the
- * applicable Container's ACL is not accessible to the authenticated user, `acl.fallbackAcl` will be
- * `null`.
- *
- * @param url URL of the LitDataset to fetch.
- * @param options Optional parameter `options.fetch`: An alternative `fetch` function to make the HTTP request, compatible with the browser-native [fetch API](https://developer.mozilla.org/docs/Web/API/WindowOrWorkerGlobalScope/fetch#parameters).
- * @returns A Resource's metadata and the ACLs that apply to the Resource, if available to the authenticated user.
- */
-export async function unstable_fetchResourceInfoWithAcl(
-  url: UrlString,
-  options: Partial<typeof defaultFetchOptions> = defaultFetchOptions
-): Promise<WithResourceInfo & unstable_WithAcl> {
-  const resourceInfo = await internal_fetchResourceInfo(url, options);
-  const acl = await internal_fetchAcl({ resourceInfo }, options);
-  return Object.assign({ resourceInfo }, { acl });
-}
-
-/**
- * @internal
- */
-export function parseResourceInfo(
-  response: Response
-): WithResourceInfo["resourceInfo"] {
-  const contentTypeParts =
-    response.headers.get("Content-Type")?.split(";") ?? [];
-  const isLitDataset =
-    contentTypeParts.length > 0 &&
-    ["text/turtle", "application/ld+json"].includes(contentTypeParts[0]);
-
-  const resourceInfo: WithResourceInfo["resourceInfo"] = {
-    fetchedFrom: response.url,
-    isLitDataset: isLitDataset,
-    contentType: response.headers.get("Content-Type") ?? undefined,
-  };
-
-  const linkHeader = response.headers.get("Link");
-  if (linkHeader) {
-    const parsedLinks = LinkHeader.parse(linkHeader);
-    const aclLinks = parsedLinks.get("rel", "acl");
-    if (aclLinks.length === 1) {
-      resourceInfo.unstable_aclUrl = new URL(
-        aclLinks[0].uri,
-        resourceInfo.fetchedFrom
-      ).href;
-    }
-  }
-
-  const wacAllowHeader = response.headers.get("WAC-Allow");
-  if (wacAllowHeader) {
-    resourceInfo.unstable_permissions = parseWacAllowHeader(wacAllowHeader);
-  }
-
-  return resourceInfo;
-}
-
-/**
- * @param resource Resource for which to check whether it is a Container.
- * @returns Whether `resource` is a Container.
- */
-export function isContainer(resource: WithResourceInfo): boolean {
-  return getFetchedFrom(resource).endsWith("/");
-}
-
-/**
- * @param resource Resource for which to check whether it contains a LitDataset.
- * @return Whether `resource` contains a LitDataset.
- */
-export function isLitDataset(resource: WithResourceInfo): boolean {
-  return resource.resourceInfo.isLitDataset;
-}
-
-/**
- * @param resource Resource for which to determine the Content Type.
- * @returns The Content Type, if known, or null if not known.
- */
-export function getContentType(resource: WithResourceInfo): string | null {
-  return resource.resourceInfo.contentType ?? null;
-}
-
-/**
- * @param resource
- * @returns The URL from which the resource has been fetched
- */
-export function getFetchedFrom(resource: WithResourceInfo): string {
-  return resource.resourceInfo.fetchedFrom;
 }
 
 /**
@@ -267,16 +107,15 @@ export function getFetchedFrom(resource: WithResourceInfo): string {
  */
 export async function unstable_fetchLitDatasetWithAcl(
   url: UrlString,
-  options: Partial<typeof defaultFetchOptions> = defaultFetchOptions
+  options: Partial<
+    typeof internal_defaultFetchOptions
+  > = internal_defaultFetchOptions
 ): Promise<LitDataset & WithResourceInfo & unstable_WithAcl> {
   const litDataset = await fetchLitDataset(url, options);
   const acl = await internal_fetchAcl(litDataset, options);
   return Object.assign(litDataset, { acl });
 }
 
-const defaultSaveOptions = {
-  fetch: fetch,
-};
 /**
  * Given a LitDataset, store it in a Solid Pod (overwriting the existing data at the given URL).
  *
@@ -288,10 +127,12 @@ const defaultSaveOptions = {
 export async function saveLitDatasetAt(
   url: UrlString,
   litDataset: LitDataset,
-  options: Partial<typeof defaultSaveOptions> = defaultSaveOptions
+  options: Partial<
+    typeof internal_defaultFetchOptions
+  > = internal_defaultFetchOptions
 ): Promise<LitDataset & WithResourceInfo & WithChangeLog> {
   const config = {
-    ...defaultSaveOptions,
+    ...internal_defaultFetchOptions,
     ...options,
   };
 
@@ -377,11 +218,8 @@ function isUpdate(
   );
 }
 
-const defaultSaveInContainerOptions = {
-  fetch: fetch,
-};
 type SaveInContainerOptions = Partial<
-  typeof defaultSaveInContainerOptions & {
+  typeof internal_defaultFetchOptions & {
     slugSuggestion: string;
   }
 >;
@@ -396,10 +234,10 @@ type SaveInContainerOptions = Partial<
 export async function saveLitDatasetInContainer(
   containerUrl: UrlString,
   litDataset: LitDataset,
-  options: SaveInContainerOptions = defaultSaveInContainerOptions
+  options: SaveInContainerOptions = internal_defaultFetchOptions
 ): Promise<LitDataset & WithResourceInfo> {
   const config = {
-    ...defaultSaveOptions,
+    ...internal_defaultFetchOptions,
     ...options,
   };
 
@@ -452,67 +290,6 @@ export async function saveLitDatasetInContainer(
   return resourceWithResolvedIris;
 }
 
-/**
- * Save the ACL for a Resource.
- *
- * @param resource The Resource to which the given ACL applies.
- * @param resourceAcl An [[unstable_AclDataset]] whose ACL Rules will apply to `resource`.
- * @param options Optional parameter `options.fetch`: An alternative `fetch` function to make the HTTP request, compatible with the browser-native [fetch API](https://developer.mozilla.org/docs/Web/API/WindowOrWorkerGlobalScope/fetch#parameters).
- */
-export async function unstable_saveAclFor(
-  resource: unstable_WithAccessibleAcl,
-  resourceAcl: unstable_AclDataset,
-  options: Partial<typeof defaultSaveOptions> = defaultSaveOptions
-): Promise<unstable_AclDataset & WithResourceInfo> {
-  const savedDataset = await saveLitDatasetAt(
-    resource.resourceInfo.unstable_aclUrl,
-    resourceAcl,
-    options
-  );
-  const savedAclDataset: unstable_AclDataset &
-    typeof savedDataset = Object.assign(savedDataset, {
-    accessTo: getFetchedFrom(resource),
-  });
-
-  return savedAclDataset;
-}
-
-/**
- * Remove the ACL of a Resource.
- *
- * @param resource The Resource for which you want to delete the Access Control List Resource.
- * @param options Optional parameter `options.fetch`: An alternative `fetch` function to make the HTTP request, compatible with the browser-native [fetch API](https://developer.mozilla.org/docs/Web/API/WindowOrWorkerGlobalScope/fetch#parameters).
- */
-export async function unstable_deleteAclFor<
-  Resource extends WithResourceInfo & unstable_WithAccessibleAcl
->(
-  resource: Resource,
-  options: Partial<typeof defaultSaveOptions> = defaultSaveOptions
-): Promise<Resource & { acl: { resourceAcl: null } }> {
-  const config = {
-    ...defaultSaveOptions,
-    ...options,
-  };
-
-  const response = await config.fetch(resource.resourceInfo.unstable_aclUrl, {
-    method: "DELETE",
-  });
-
-  if (!response.ok) {
-    throw new Error(
-      `Deleting the ACL failed: ${response.status} ${response.statusText}.`
-    );
-  }
-
-  const storedResource = Object.assign(resource, {
-    acl: {
-      resourceAcl: null,
-    },
-  });
-
-  return storedResource;
-}
-
 function getNamedNodesForLocalNodes(quad: Quad): Quad {
   const subject = isLocalNode(quad.subject)
     ? getNamedNodeFromLocalNode(quad.subject)
@@ -545,59 +322,4 @@ function resolveLocalIrisInLitDataset<
   });
 
   return litDataset;
-}
-
-/**
- * Parse a WAC-Allow header into user and public access booleans.
- *
- * @param wacAllowHeader A WAC-Allow header in the format `user="read append write control",public="read"`
- * @see https://github.com/solid/solid-spec/blob/cb1373a369398d561b909009bd0e5a8c3fec953b/api-rest.md#wac-allow-headers
- */
-function parseWacAllowHeader(wacAllowHeader: string) {
-  function parsePermissionStatement(
-    permissionStatement: string
-  ): unstable_Access {
-    const permissions = permissionStatement.split(" ");
-    const writePermission = permissions.includes("write");
-    return writePermission
-      ? {
-          read: permissions.includes("read"),
-          append: true,
-          write: true,
-          control: permissions.includes("control"),
-        }
-      : {
-          read: permissions.includes("read"),
-          append: permissions.includes("append"),
-          write: false,
-          control: permissions.includes("control"),
-        };
-  }
-  function getStatementFor(header: string, scope: "user" | "public") {
-    const relevantEntries = header
-      .split(",")
-      .map((rawEntry) => rawEntry.split("="))
-      .filter((parts) => parts.length === 2 && parts[0].trim() === scope);
-
-    // There should only be one statement with the given scope:
-    if (relevantEntries.length !== 1) {
-      return "";
-    }
-    const relevantStatement = relevantEntries[0][1].trim();
-
-    // The given statement should be wrapped in double quotes to be valid:
-    if (
-      relevantStatement.charAt(0) !== '"' ||
-      relevantStatement.charAt(relevantStatement.length - 1) !== '"'
-    ) {
-      return "";
-    }
-    // Return the statment without the wrapping quotes, e.g.: read append write control
-    return relevantStatement.substring(1, relevantStatement.length - 1);
-  }
-
-  return {
-    user: parsePermissionStatement(getStatementFor(wacAllowHeader, "user")),
-    public: parsePermissionStatement(getStatementFor(wacAllowHeader, "public")),
-  };
 }
